refactor(schedule): move mobile components into mobile/ subtree (Phase 3b)
Pure move of the mobile schedule-editor components out of the 1690-line
MobileScheduleView.tsx into the mobile/ subtree begun in Phase 3a. No
behavior change: symbols, JSX, styles, and comments are preserved verbatim;
only file locations and relative import paths changed.

New files (all under src/components/admin/schedule/mobile/):
- rail.ts            ← verbatim move of ../mobileRail.ts (buildTrackRail,
                       RailSegment, helpers); old file deleted
- BottomSheet.tsx    ← BottomSheet (focus trap + scroll lock)
- TrackRail.tsx      ← TrackRail + RailBody
- chrome.tsx         ← DaySelect, LegendDisclosure, DurationChip,
                       PlacingBanner, dayLabel
- sheets/ServiceEditSheet.tsx
- sheets/UnassignedDrawer.tsx
- sheets/CardActionSheet.tsx
- sheets/TrackActionSheet.tsx
- MobileScheduleView.tsx ← remaining main component (state + handlers +
                          layout); old top-level file deleted

placement.ts / railGeometry.ts now import RailSegment from ./rail. The
mobile/ barrel re-exports rail + all the moved components (MobileScheduleView
is the only one external code needs today). Importers updated to pull
MobileScheduleView / RailSegment / buildTrackRail through the ./mobile barrel
(ScheduleEditor, the story, and the three test files).

Verification: tsc --noEmit clean, eslint clean, prettier clean, vitest
141/141 pass, pnpm shoot OK (3 cards checked, none exceed the 393px viewport).
