+++ __tests__/components/admin/schedule/mobile-placement.test.ts
@@ -10,7 +10,7 @@
 import { describe, it, expect } from 'vitest'
 import type { ScheduleTrack, TrackTalk } from '@/lib/conference/types'
 import type { ProposalExisting } from '@/lib/proposal/types'
-import type { RailSegment } from '@/components/admin/schedule/mobileRail'
+import type { RailSegment } from '@/components/admin/schedule/mobile'
 import type { Placing } from '@/components/admin/schedule/mobile'
 import { segmentState } from '@/components/admin/schedule/mobile'
