+++ src/components/admin/schedule/ScheduleEditor.tsx
@@ -35,7 +35,7 @@ import { UnassignedProposals } from './UnassignedProposals'
 import { MemoizedDroppableTrack as DroppableTrack } from './DroppableTrack'
 import { DraggableProposal } from './DraggableProposal'
 import { DraggableServiceSession } from './DraggableServiceSession'
-import { MobileScheduleView } from './MobileScheduleView'
+import { MobileScheduleView } from './mobile'
 import { HeaderSection } from './HeaderSection'
 import { AddTrackModal } from './AddTrackModal'
 import { ScheduleProvider } from './ScheduleContext'
