+++ src/components/admin/schedule/mobile/index.ts
@@ -1,9 +1,23 @@
 // Public surface of the mobile schedule-editor subtree. Import from
 // `./mobile` (or `@/components/admin/schedule/mobile`) rather than reaching into
 // deep file paths, per the AGENTS.md barrel convention — so internal file names
-// can move without churning every call site. Phase 3b will add the sheet / rail
-// / chrome components and MobileScheduleView itself here.
+// can move without churning every call site.
 export * from './constants'
 export * from './types'
+export * from './rail'
 export * from './railGeometry'
 export * from './placement'
+export { BottomSheet } from './BottomSheet'
+export { TrackRail } from './TrackRail'
+export {
+  DaySelect,
+  LegendDisclosure,
+  DurationChip,
+  PlacingBanner,
+  dayLabel,
+} from './chrome'
+export { ServiceEditSheet } from './sheets/ServiceEditSheet'
+export { UnassignedDrawer } from './sheets/UnassignedDrawer'
+export { CardActionSheet } from './sheets/CardActionSheet'
+export { TrackActionSheet } from './sheets/TrackActionSheet'
+export { MobileScheduleView } from './MobileScheduleView'
