+++ src/components/admin/schedule/mobile/rail.ts
